+++ test/test.rs
@@ -1,4 +1,6 @@
 #[macro_use]
+extern crate cfg_if;
+#[macro_use]
 extern crate nix;
 #[macro_use]
 extern crate lazy_static;